requirements.txt
asgiref==3.8.1
cachetools==5.3.3
chardet==5.2.0
colorama==0.4.6
coverage==7.6.0
distlib==0.3.8
Django==4.2.9
django-environ==0.11.2
djangorestframework==3.15.1
filelock==3.15.4
mysqlclient==2.2.4
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pyproject-api==1.7.1
sqlparse==0.5.0
tomli==2.0.1
tox==4.16.0
typing_extensions==4.12.2
virtualenv==20.26.3
